Rename package service to services
