+++ test.sh
@@ -36,8 +36,8 @@ go test -v ./engine -race || exit 1
 go test -i ./messengers || exit 1
 go test -v ./messengers -race || exit 1
 
-go test -i ./service || exit 1
-go test -v ./service -race || exit 1
+go test -i ./services || exit 1
+go test -v ./services -race || exit 1
 
 #go test -i ./tests/functional
 #ETCD_BIN_PATH=$(pwd)/bin/walter go test -v ./tests/functional -race
